+++ packages/bloc_test/CHANGELOG.md
@@ -1,3 +1,16 @@
+# 8.0.0-nullsafety.4
+
+- **BREAKING**: feat: `seed` returns a `Function` to support dynamic seed values
+
+# 8.0.0-nullsafety.3
+
+- feat: upgrade to `mocktail: ">=0.0.2-dev.5 <0.0.2"`
+
+# 8.0.0-nullsafety.2
+
+- fix: restrict to `mocktail: ">=0.0.1-dev.12 <0.0.1"`
+- feat: use `package:test_api` instead of `package:test` for sound null safety
+
 # 8.0.0-nullsafety.1
 
 - chore: upgrade to `bloc ^7.0.0-nullsafety.3`